+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ---------------------
 ##### Fixed
 - Timepicker and Datetime Picker issues around empty and default values.
+- When loading a record with nested repeats the second+ series gets inserted out-of-order.
 
 [4.34.0] - 2017-05-22
 ---------------------